requirements.txt
naff
git+https://github.com/KAJdev/async-openai
git+https://github.com/KAJdev/dash
aiohttp
motor
pymongo[srv]
humanize
python-dotenv
orjson
dacite
pytesseract
Pillow
bs4